+++ backend/swagger/docs/docs.go
@@ -4102,6 +4102,9 @@ const docTemplate = `{
                 "bannerUrl": {
                     "type": "string"
                 },
+                "canEditShare": {
+                    "type": "boolean"
+                },
                 "description": {
                     "type": "string"
                 },
@@ -5466,6 +5469,10 @@ const docTemplate = `{
                     "description": "debug onlyoffice editor",
                     "type": "boolean"
                 },
+                "deleteAfterArchive": {
+                    "description": "delete source files after successful creation/extraction of archives",
+                    "type": "boolean"
+                },
                 "deleteWithoutConfirming": {
                     "description": "delete files without confirmation",
                     "type": "boolean"
@@ -5517,6 +5524,10 @@ const docTemplate = `{
                 "gallerySize": {
                     "description": "0-9 - the size of the gallery thumbnails",
                     "type": "integer"
+                },
+                "hideFilesInTree": {
+                    "description": "hide files in the sidebar tree navigation, when true, will show only directories.",
+                    "type": "boolean"
                 },
                 "hideSidebarFileActions": {
                     "description": "hide the file actions in the sidebar",
@@ -5610,6 +5621,10 @@ const docTemplate = `{
         "settings.UserDefaultsPreview": {
             "type": "object",
             "properties": {
+                "audio": {
+                    "description": "show thumbnails for audio files",
+                    "type": "boolean"
+                },
                 "autoplayMedia": {
                     "description": "autoplay media files in preview",
                     "type": "boolean"
@@ -5632,6 +5647,10 @@ const docTemplate = `{
                 },
                 "image": {
                     "description": "show thumbnails for image files",
+                    "type": "boolean"
+                },
+                "models": {
+                    "description": "show live thumbnails for 3D models files",
                     "type": "boolean"
                 },
                 "motionVideoPreview": {
@@ -5681,6 +5700,9 @@ const docTemplate = `{
                 },
                 "bannerUrl": {
                     "type": "string"
+                },
+                "canEditShare": {
+                    "type": "boolean"
                 },
                 "description": {
                     "type": "string"
@@ -5822,6 +5844,9 @@ const docTemplate = `{
                 },
                 "bannerUrl": {
                     "type": "string"
+                },
+                "canEditShare": {
+                    "type": "boolean"
                 },
                 "description": {
                     "type": "string"
@@ -5975,6 +6000,9 @@ const docTemplate = `{
                 },
                 "bannerUrl": {
                     "type": "string"
+                },
+                "canEditShare": {
+                    "type": "boolean"
                 },
                 "description": {
                     "type": "string"
@@ -6217,6 +6245,10 @@ const docTemplate = `{
         "users.Preview": {
             "type": "object",
             "properties": {
+                "audio": {
+                    "description": "show thumbnail preview image for audio files",
+                    "type": "boolean"
+                },
                 "autoplayMedia": {
                     "description": "autoplay media files in preview",
                     "type": "boolean"
@@ -6235,6 +6267,10 @@ const docTemplate = `{
                 },
                 "image": {
                     "description": "show thumbnail preview image for image files",
+                    "type": "boolean"
+                },
+                "models": {
+                    "description": "show live thumbnails for 3D models files",
                     "type": "boolean"
                 },
                 "motionVideoPreview": {
@@ -6328,6 +6364,10 @@ const docTemplate = `{
                     "description": "debug onlyoffice editor",
                     "type": "boolean"
                 },
+                "deleteAfterArchive": {
+                    "description": "delete source files after successful creation/extraction of archives",
+                    "type": "boolean"
+                },
                 "deleteWithoutConfirming": {
                     "description": "delete files without confirmation",
                     "type": "boolean"
@@ -6378,6 +6418,10 @@ const docTemplate = `{
                 "gallerySize": {
                     "description": "0-9 - the size of the gallery thumbnails",
                     "type": "integer"
+                },
+                "hideFilesInTree": {
+                    "description": "hide files in the sidebar tree navigation, when true, will show only directories.",
+                    "type": "boolean"
                 },
                 "hideSidebarFileActions": {
                     "description": "hide the file actions in the sidebar",