+++ frontend/src/components/files/PopupPreview.vue
@@ -41,6 +41,9 @@ export default {
   },
   computed: {
     sourceInfo() {
+      if (state.multiple) {
+        return false;
+      }
       return state.popupPreviewSourceInfo;
     },
   },